[MRELEASE-159] Support a pattern to generate the release tag.
o I had a quick look at the suggested solutions and basing off of MessageFormat requires people to remember exactly which {0} and {1} correspond with. This solution uses plexus-interpolation to interpolate properties delimited with @{ and } so that the properties escape Maven's default interpolation which would resolve them in the -SNAPSHOT project.

o May need to double check that the site generates correctly

git-svn-id: https://svn.apache.org/repos/asf/maven/release/trunk@1139359 13f79535-47bb-0310-9956-ffa450edef68