Add a solution for running org.apache.maven.shared.release.exec.InvokerMavenExecutorTest.
The Maven shell scripts do not export M2_HOME correctly. The system property maven.home
is not inherited by default in surefire. Directly setting maven.home here does not work
because a lot of the tests are checking M2_HOME.
So introduce a system property injectedMavenHome in surefire which is picked up by
InvokerMavenExecutorTest lateron to determine maven.home. 


git-svn-id: https://svn.apache.org/repos/asf/maven/release/trunk@1669230 13f79535-47bb-0310-9956-ffa450edef68