[MRELEASE-964] Error injecting: org.apache.maven.shared.release.phase.RewritePomsForReleasePhase
jaxen is bringing a conflicting dependency of JDom (version 1.0 when the plugin depends on 1.1). Instead of excluding jdom, the Jaxen dependency can be removed as XPath is not used anymore: it was introduced in r397561 for the RewritePomsForReleasePhase class, the code depending on it was later moved to AbstractRewritePomsPhase in r398224, and that code was rewritten in r1179704, removing the XPath dependency.

git-svn-id: https://svn.apache.org/repos/asf/maven/release/trunk@1777913 13f79535-47bb-0310-9956-ffa450edef68