[MRELEASE-975] NPE when using an unknown project versionpolicy id
Passing an empty version policy hint (with -DprojectVersionPolicyId on the command-line), on an unknown one, should result in a helpful exception being thrown (not NullPointerException). Upgrading JUnit to 4.11 to use the expected exception mechanism (in order to test that the cause of the exception thrown is a PolicyException). This closes #8.

git-svn-id: https://svn.apache.org/repos/asf/maven/release/trunk@1777917 13f79535-47bb-0310-9956-ffa450edef68