[MRELEASE-976] release:branch should also support project version policies
Adding parameter projectVersionPolicyId to the branch and update-versions goals. The ITs are done by creating a stub version policy that switches between snapshot and release and testing that it is possible to invoke release:branch, release:update-versions and release:prepare on a sample project with this custom policy. This closes #9.

git-svn-id: https://svn.apache.org/repos/asf/maven/release/trunk@1777925 13f79535-47bb-0310-9956-ffa450edef68